requirements.txt

flask
flask-cors
langchain
langchain-google-genai
python-dotenv
gunicorn
